Php 8.0

+++ php8/ext/mysqli/tests/mysqli_debug.phpt
@@ -1,9 +1,8 @@
 --TEST--
 mysqli_debug()
---EXTENSIONS--
-mysqli
 --SKIPIF--
 <?php
+require_once('skipif.inc');
 require_once('skipifconnectfailure.inc');
 
 if (!function_exists('mysqli_debug'))
@@ -23,34 +22,38 @@ if (defined('MYSQLI_DEBUG_TRACE_ENABLED') && !MYSQLI_DEBUG_TRACE_ENABLED)
     if (true !== ($tmp = mysqli_debug(sprintf('d:t:O,%s/mysqli_debug_phpt.trace', sys_get_temp_dir()))))
         printf("[002] Expecting boolean/true, got %s/%s\n", gettype($tmp), $tmp);
 
-    // table.inc will create a database connection and run some SQL queries, therefore
-    // the debug file should have entries
-    require_once('table.inc');
+    if ($IS_MYSQLND) {
+        // let's make this mysqlnd only - for libmysql we need debug installation
 
-    clearstatcache();
-    $trace_file = sprintf('%s/mysqli_debug_phpt.trace', sys_get_temp_dir());
-    if (!file_exists($trace_file))
-        printf("[003] Trace file '%s' has not been created\n", $trace_file);
-    if (filesize($trace_file) < 50)
-        printf("[004] Trace file '%s' is very small. filesize() reports only %d bytes. Please check.\n",
-            $trace_file,
-            filesize($trace_file));
+        // table.inc will create a database connection and run some SQL queries, therefore
+        // the debug file should have entries
+        require_once('table.inc');
 
-    // will mysqli_debug() mind if the trace file gets removed?
-    unlink($trace_file);
-    clearstatcache();
+        clearstatcache();
+        $trace_file = sprintf('%s/mysqli_debug_phpt.trace', sys_get_temp_dir());
+        if (!file_exists($trace_file))
+            printf("[003] Trace file '%s' has not been created\n", $trace_file);
+        if (filesize($trace_file) < 50)
+            printf("[004] Trace file '%s' is very small. filesize() reports only %d bytes. Please check.\n",
+                $trace_file,
+                filesize($trace_file));
 
-    if (!$res = mysqli_query($link, 'SELECT * FROM test'))
-            printf("[005] [%d] %s\n", mysqli_errno($link), mysqli_error($link));
-        else
-            mysqli_free_result($res);
+        // will mysqli_debug() mind if the trace file gets removed?
+        unlink($trace_file);
+        clearstatcache();
 
-    mysqli_close($link);
+        if (!$res = mysqli_query($link, 'SELECT * FROM test'))
+                printf("[005] [%d] %s\n", mysqli_errno($link), mysqli_error($link));
+            else
+                mysqli_free_result($res);
 
-    clearstatcache();
-    if (!file_exists($trace_file))
-        printf("[006] Trace file '%s' does not exist\n", $trace_file);
-    unlink($trace_file);
+        mysqli_close($link);
+
+        clearstatcache();
+        if (!file_exists($trace_file))
+            printf("[006] Trace file '%s' does not exist\n", $trace_file);
+        unlink($trace_file);
+    }
 
     print "done!";
 ?>
